requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
dataclasses-json==0.6.7
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.139
lxml==5.3.0
lxml_html_clean==0.3.1
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
ollama==0.3.3
orjson==3.10.11
packaging==24.1
pgvector==0.3.6
pillow==11.0.0
propcache==0.2.0
psycopg==3.2.3
psycopg2==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==75.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tenacity==9.0.0
timescale-vector==0.0.7
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.1
Werkzeug==3.1.2
wheel==0.44.0
yarl==1.17.1
